feat(trouble.lua): add trouble.nvim plugin with configuration to the project
The trouble.nvim plugin is added to the project. It is a plugin that provides a better way to handle and navigate through the list of diagnostics and other code issues. The plugin is added as a dependency and requires the nvim-tree/nvim-web-devicons plugin to function properly. The configuration for the trouble.nvim plugin is added in the absolute.after.trouble.lua file.