feat(nvim-surround.lua): add nvim-surround plugin
Add the nvim-surround plugin to the list of plugins used in the configuration. This plugin provides convenient mappings and commands for manipulating surrounding characters. The plugin is sourced from the "kylechui/nvim-surround" repository, using the latest version available. The plugin is configured to be loaded during the "VeryLazy" event, and the necessary configuration file is required to be loaded after the plugin is initialized.