feat(tailwindcss-colors.lua): add support for tailwindcss-colors plugin
The `tailwindcss-colors.lua` file was added to the `absolute/after` directory. This file imports the `tailwindcss-colors` module and sets it up.

The `tailwindcss-colors.lua` file was also added to the `absolute/plugins` directory. This file returns a table with the plugin configuration, including a `config` function that requires the `absolute.after.tailwindcss-colors` module.

These changes were made to add support for the `tailwindcss-colors` plugin in the Absolute Neovim configuration.